Add onlogin add shoppinglist
