+++ imports/ui/actions/AppActions.js
@@ -97,6 +97,7 @@ export const addNewShoppngList = item => {
         dispatch(fetchItemsFailure(err));
       } else {
         //Callback "res" is the ID of the successfully added item
+        console.log(res);
         item._id = res;
         dispatch(addNewShoppingList(item));
       }
